Update Backlog Controller

+++ src/controllers/BacklogController.ts
@@ -70,24 +70,78 @@ export class BacklogController extends BaseController {
     return null;
   }
 
+  @Response<ForbiddenError>(403, 'Access Denied. You are not authorized to perform this action.')
+  @Response<UnauthorizedError>(401, 'Unauthorized user.')
   @Response<MenuchiError>(400, 'All item IDs must be in the request.')
   @SuccessResponse(204, 'Item orders in the category updated successfully.')
+  @Security('', [RolesEnum.RestaurantOwner])
   @Patch('/{backlogId}/reorder-items/in-category')
-  public async reorderItemsInCategory(@Path() backlogId: UUID, @Body() body: UUID[]): Promise<number> {
+  public async reorderItemsInCategory(
+    @Path() backlogId: UUID,
+    @Body() body: UUID[],
+    @Request() req: express.Request
+  ): Promise<number> {
+    this.checkPermission(req.session.user, PermissionScope.Backlog, backlogId);
     return BacklogService.reorderItemsInCategory(backlogId, body);
   }
 
+  @Response<ForbiddenError>(403, 'Access Denied. You are not authorized to perform this action.')
+  @Response<UnauthorizedError>(401, 'Unauthorized user.')
   @Response<MenuchiError>(400, 'All item IDs must be in the request.')
   @SuccessResponse(204, 'Item orders in the list updated successfully.')
+  @Security('', [RolesEnum.RestaurantOwner])
   @Patch('/{backlogId}/reorder-items/in-list')
-  public async reorderItemsInList(@Path() backlogId: UUID, @Body() body: UUID[]): Promise<number> {
+  public async reorderItemsInList(
+    @Path() backlogId: UUID,
+    @Body() body: UUID[],
+    @Request() req: express.Request
+  ): Promise<number> {
+    this.checkPermission(req.session.user, PermissionScope.Backlog, backlogId);
     return BacklogService.reorderItemsInList(backlogId, body);
   }
 
+  @Response<ForbiddenError>(403, 'Access Denied. You are not authorized to perform this action.')
+  @Response<UnauthorizedError>(401, 'Unauthorized user.')
   @SuccessResponse(204, 'Items deleted successfully. It doesn\'t retrieve anything.')
+  @Security('', [RolesEnum.RestaurantOwner])
   @Delete('/{backlogId}/items')
-  public async deleteItems(@Path() backlogId: UUID, @Body() body: UUID[]): Promise<null> {
+  public async deleteItems(
+    @Path() backlogId: UUID,
+    @Body() body: UUID[],
+    @Request() req: express.Request
+  ): Promise<null> {
+    this.checkPermission(req.session.user, PermissionScope.Backlog, backlogId);
     await BacklogService.deleteItems(backlogId, body);
+    return null;
+  }
+
+  @Response<ForbiddenError>(403, 'Access Denied. You are not authorized to perform this action.')
+  @Response<UnauthorizedError>(401, 'Unauthorized user.')
+  @Response<MenuchiError>(400, 'All category IDs must be in the request.')
+  @Security('', [RolesEnum.RestaurantOwner])
+  @SuccessResponse(204, 'Category orders in the backlog updated successfully.')
+  @Patch('/{backlogId}/reorder-categories')
+  public async reorderCategoriesInBacklog(
+    @Path() backlogId: UUID,
+    @Body() body: UUID[],
+    @Request() req: express.Request
+  ): Promise<number> {
+    this.checkPermission(req.session.user, PermissionScope.Backlog, backlogId);
+    return BacklogService.reorderCategoriesInBacklog(backlogId, body);
+  }
+
+  @Response<ForbiddenError>(403, 'Access Denied. You are not authorized to perform this action.')
+  @Response<UnauthorizedError>(401, 'Unauthorized user.')
+  @SuccessResponse(204, 'Category and its items deleted successfully.  It doesn\'t retrieve anything.')
+  @Security('', [RolesEnum.RestaurantOwner])
+  @Delete('/{backlogId}/categories/{categoryId}')
+  public async deleteCategory(
+    @Path() backlogId: UUID,
+    @Path() categoryId: UUID,
+    @Request() req: express.Request
+  ): Promise<null> {
+    this.checkPermission(req.session.user, PermissionScope.Backlog, backlogId);
+    await BacklogService.deleteCategory(backlogId, categoryId);
     return null;
   }
 }
